requirements.txt
PyPDF2==3.0.1
openai==0.8.0
sklearn-pandas==2.2.0